Added WebSocketController test and restructured rest tests directory tree. This commit refs #2641 @0h36m
