+++ src/test/java/es/udc/fi/dc/fd/rest/controllers/WebSocketControllerTest.java
@@ -1,0 +1,35 @@
+package es.udc.fi.dc.fd.rest.controllers;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringRunner;
+import org.springframework.transaction.annotation.Transactional;
+
+import es.udc.fi.dc.fd.rest.dtos.PostStreamDto;
+
+
+
+@RunWith(SpringRunner.class)
+@SpringBootTest
+@AutoConfigureMockMvc
+@ActiveProfiles("test")
+@Transactional
+public class WebSocketControllerTest {
+
+    @Autowired
+    WebSocketController webSocketController;
+
+    @Test
+    public void notifyNewPostTest() {
+        PostStreamDto postStreamDto = new PostStreamDto("posts.newPost");
+
+        assertEquals(postStreamDto, webSocketController.notifyNewPost(postStreamDto));
+    }
+    
+}
