+++ src/test/java/es/udc/fi/dc/fd/rest/dtos/BlockDtoTest.java
@@ -1,4 +1,4 @@
-package es.udc.fi.dc.fd.rest;
+package es.udc.fi.dc.fd.rest.dtos;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -11,8 +11,6 @@ import java.util.Arrays;
 import java.util.List;
 
 import org.junit.jupiter.api.Test;
-
-import es.udc.fi.dc.fd.rest.dtos.BlockDto;
 
 public class BlockDtoTest {
