+++ src/test/java/es/udc/fi/dc/fd/rest/dtos/CategoryConversorTest.java
@@ -1,4 +1,4 @@
-package es.udc.fi.dc.fd.rest;
+package es.udc.fi.dc.fd.rest.dtos;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -9,8 +9,6 @@ import java.util.List;
 import org.junit.jupiter.api.Test;
 
 import es.udc.fi.dc.fd.model.entities.Category;
-import es.udc.fi.dc.fd.rest.dtos.CategoryConversor;
-import es.udc.fi.dc.fd.rest.dtos.CategoryDto;
 
 public class CategoryConversorTest {
